requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
